% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/checkRequired.R
\name{checkRequired}
\alias{checkRequired}
\title{checks for required fields in a response
TODO make this function only usable within library}
\usage{
checkRequired(var, obj)
}
\arguments{
\item{var}{the required variable to check}

\item{obj}{the parent function of checkRequired}
}
\description{
checks for required fields in a response
TODO make this function only usable within library
}

