% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/GASearchVariantsRequest.R
\name{GASearchVariantsRequest}
\alias{GASearchVariantsRequest}
\title{Constructs a GASearchVariantsRequest
TODO add params}
\usage{
GASearchVariantsRequest(variantSetIds, variantName = NULL,
  callSetIds = NULL, referenceName, start = NULL, end, pageSize = NULL,
  pageToken = NULL)
}
\description{
Constructs a GASearchVariantsRequest
TODO add params
}

