% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/variant_search.R
\name{variant_search}
\alias{variant_search}
\title{POST /variants/search}
\usage{
variant_search(request)
}
\arguments{
\item{request}{A GASearchVariantsRequest for POST}
}
\value{
response A GASearchVariantsResponse
}
\description{
Query to return calls associated with a variantset
}

